control de acceso login y skp

+++ src/main/java/es/uji/ei1027/skillSharing/controller/UserController.java
@@ -2,6 +2,7 @@ package es.uji.ei1027.skillSharing.controller;
 
 import es.uji.ei1027.skillSharing.dao.UsuarioDao;
 import es.uji.ei1027.skillSharing.modelo.Usuario;
+import org.apache.catalina.User;
 import org.jasypt.util.password.BasicPasswordEncryptor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import javax.servlet.http.HttpSession;
 import java.util.Arrays;
 import java.util.List;
 
@@ -25,12 +27,25 @@ public class UserController {
         this.usuarioDao=usuarioDao;
     }
     @RequestMapping("/list")
-    public String listUsuarios(Model model) {
+    public String listUsuarios(HttpSession session, Model model) {
+        if (session.getAttribute("user") == null){
+            session.setAttribute("nextUrl","/usuario/list");
+            model.addAttribute("usuario", new Usuario());
+            return "login";
+        }
         model.addAttribute("usuarios", usuarioDao.getUsuarios());
         return "usuario/list";
     }
     @RequestMapping(value="/add")
-    public String addUsuario(Model model) {
+    public String addUsuario(HttpSession session,Model model) {
+        if (session.getAttribute("user") == null){
+            session.setAttribute("nextUrl","/usuario/list");
+            return "login";
+        }
+        Usuario user = (Usuario)session.getAttribute("user");
+        if (!user.isSkp()){
+            return "forbiden";
+        }
         model.addAttribute("usuario", new Usuario());
         return "usuario/add";
     }
@@ -48,7 +63,15 @@ public class UserController {
         return "redirect:list";
     }
     @RequestMapping(value="/update/{username}", method = RequestMethod.GET)
-    public String editUsuario(Model model, @PathVariable String username) {
+    public String editUsuario(HttpSession session, Model model, @PathVariable String username) {
+        if (session.getAttribute("user") == null){
+            session.setAttribute("nextUrl","/usuario/list");
+            return "login";
+        }
+        Usuario user = (Usuario)session.getAttribute("user");
+        if (!user.isSkp()){
+            return "forbiden";
+        }
         model.addAttribute("usuario", usuarioDao.getUsuario(username));
         return "usuario/update";
     }
@@ -61,12 +84,28 @@ public class UserController {
         return "redirect:list";
     }
     @RequestMapping(value="/delete/{username}")
-    public String processDelete(@PathVariable String username) {
+    public String processDelete(HttpSession session, @PathVariable String username) {
+        if (session.getAttribute("user") == null){
+            session.setAttribute("nextUrl","/usuario/list");
+            return "login";
+        }
+        Usuario user = (Usuario)session.getAttribute("user");
+        if (!user.isSkp()){
+            return "forbiden";
+        }
         usuarioDao.deleteUsuario(username);
         return "redirect:../list";
     }
     @RequestMapping(value="/passwd/{username}", method = RequestMethod.GET)
-    public String editPassword(Model model, @PathVariable String username) {
+    public String editPassword(HttpSession session, Model model, @PathVariable String username) {
+        if (session.getAttribute("user") == null){
+            session.setAttribute("nextUrl","/usuario/list");
+            return "login";
+        }
+        Usuario user = (Usuario)session.getAttribute("user");
+        if (!user.isSkp()){
+            return "forbiden";
+        }
         model.addAttribute("usuario", usuarioDao.getUsuario(username));
         return "usuario/passwd";
     }
